refactor: trackArtistResponse dto 위치 이동
