+++ src/main/java/com/team/comma/domain/favorite/track/repository/FavoriteTrackRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.team.comma.domain.favorite.track.repository;
 
 import com.team.comma.domain.favorite.track.dto.FavoriteTrackResponse;
-import com.team.comma.domain.track.track.dto.TrackArtistResponse;
+import com.team.comma.domain.track.artist.dto.TrackArtistResponse;
 import com.team.comma.domain.user.user.domain.User;
 
 import java.util.List;
